# goland_archpath/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для системы управления артефактами и заявками",
        "title": "ArchPath API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/artifacts": {
            "get": {
                "description": "Get all artifacts with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "artifacts"
                ],
                "summary": "Get all artifacts",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Filter by active status",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by production center",
                        "name": "production_center",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/artifact.Artifact"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve artifacts",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Create a new artifact (moderator only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "artifacts"
                ],
                "summary": "Create artifact",
                "parameters": [
                    {
                        "description": "Artifact data",
                        "name": "artifact",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/artifact.Artifact"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/artifact.Artifact"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden: moderator access required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Creation failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/artifacts/{id}": {
            "get": {
                "description": "Get a single artifact by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "artifacts"
                ],
                "summary": "Get artifact by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Artifact ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/artifact.Artifact"
                        }
                    },
                    "404": {
                        "description": "Artifact not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Update an existing artifact (moderator only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "artifacts"
                ],
                "summary": "Update artifact",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Artifact ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Artifact data",
                        "name": "artifact",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/artifact.Artifact"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden: moderator access required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Update failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Delete an artifact (moderator only)",
                "tags": [
                    "artifacts"
                ],
                "summary": "Delete artifact",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Artifact ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden: moderator access required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Deletion failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/artifacts/{id}/add-to-analysis": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Add an artifact to the user's draft trade analysis request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "artifacts"
                ],
                "summary": "Add artifact to draft request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Artifact ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Quantity and comment",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "comment": {
                                    "type": "string"
                                },
                                "quantity": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to add to draft",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/artifacts/{id}/image": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Upload an image for an artifact (moderator only)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "artifacts"
                ],
                "summary": "Upload artifact image",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Artifact ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Image file",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "image_url": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden: moderator access required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Upload failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "artifact.Artifact": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "production_center": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}